+++ Vault-Coder/src/utopia/vault/coder/util/VaultPackages.scala
@@ -28,6 +28,7 @@ object VaultPackages
 		lazy val database = base / "database"
 		lazy val models = base / "model"
 		lazy val templateModels = models/"template"
+		lazy val immutableModels = models/"immutable"
 		lazy val sql = base / "sql"
 		lazy val noSql = base / "nosql"
 		lazy val deprecation = noSql / "storable.deprecation"